HDDS-9405: replace implementation
